remotefilelog: speed up resolvefilenames
Summary:
The basestore resolvefilenames method is used to do a reverse lookup from
filename sha to filename.  Sadly this information isn't stored anywhere, so we
must search through the changelog and manifest for it.

The manifest and full changelog is too large to search in large repos,
especially when some files may not even exist there.  Only search draft commits
and the last 250,000 commits in the changelog.  If we don't find it there, we
give up.  In this case, repack will leave the file loose.

Reviewed By: DurhamG

Differential Revision: D10516853

fbshipit-source-id: 5e0e60d47b18e34ef9d692524e29b13392da8952