remotefilelog: handle corrupt pack files during repack
Summary:
If a corrupt pack file is encountered while marking the ledger for repack,
handle the failure.

Don't immediately delete the pack files.  Keep them around as we may be able
to repack some of the non-corrupt contents (particularly when the file has
been truncated).

Once repack has completed, revisit the corrupt packs that we found.  For shared
packs, delete them.  For local packs, rename them.  They are the only source of
the data, and we may be able to recover more data manually.

Reviewed By: DurhamG

Differential Revision: D10527544

fbshipit-source-id: f228fb4de3ff016a1c9035823bfa1d6b7767cdcb