hggit: don't load entire git tree at the beginning
Summary:
When doing a git import, it would load the entire git tree for the
commit into memory at once. In a large repository this can be very expensive,
especially if there are a lot of pack files.

Let's change the structure to dynamically load trees as they're requested. This
makes it O(changes imported) instead of O(trees in master).

Reviewed By: markbt

Differential Revision: D10858958

fbshipit-source-id: 1550acbd409f8b507339a2f123422ac0fbd79215