hggit: don't check for object existence when importing git objects
Summary:
When exporting a hg commit into a git commit we'd check if individual
objects existed before we actually wrote them into git. In the majority case the
object does not already exist, and for very large repositories with lots of pack
files checking this is quite expensive.

Instead, let's just write the blob into the store.  In the unlikely case that
the blob already existed, it will get de-duped during repack. This saves us
50% of time spent in some cases.

Reviewed By: markbt

Differential Revision: D10861852

fbshipit-source-id: 26718bd2e2cf40099ccf5d86e68607cd3dc9dc5a