hggit: don't scan the entire changelog during hg gexport
Summary:
When exporting hg commits to git, previously it would compute a set of
all nodes in the changelog, and all nodes in the git map, and figure out which
changelog nodes were missing. In a large repository, this is very expensive.

Instead, let's start at the heads and walk backwards.  Usually we're doing an
incremental export, so this very quickly traverse all the missing nodes and
finishes. This saves 10+ seconds off gexport.

Reviewed By: markbt

Differential Revision: D10861851

fbshipit-source-id: c7c7f45d6e60e3ee5ffa27e2d3c5fb705b41b1c1