Create test utilities for systemd service management
Summary:
In order to test 'eden start', etc. when edenfs is managed by systemd, we need to install the systemd service and let the Eden CLI start and stop the service. To avoid interfering with the user's running edenfs systemd service, and to avoid interference between unrelated tests, tests need to create a new instance of systemd.

Unfortunately, we can't create a new instance of systemd if a machine is not managed by systemd. However, we can trick systemd into thinking that the machine *is* managed by systemd. When we do that, the systemd instance seems to work without any issues.

Create a library which, when injected into systemd using LD_PRELOAD, makes systemd thinks the machine is managed by systemd.

The library will be used in a subsequent diff (D10286940). This diff is separate so it's easier to review.

Reviewed By: simpkins

Differential Revision: D10371831

fbshipit-source-id: b54b704dc631ebdc5091320dd6b680ca255f3ba2