hg: remove .watchmanconfig from !repo root
Summary:
This was added to rHG but is potentially harmful in fbsource,
leading to overlapping watches (which are exponentially expensive rather
than linear) and this empty config may cause certain features and tunings
to be disabled if you happen to watch this subdir first.
Finally, in Eden, watchman will refuse to watch this location anyway
because it is not at the repo root.

Reviewed By: singhsrb

Differential Revision: D10857366

fbshipit-source-id: 0d63ac60c6c6ca655b0d2df43d566d497520f1c2