threading: use safer Rust in threading.Condition
Summary:
Suggested by jsgf, use Condvar to emulate the "blocking" behavior so we do not
need to hold MutexGuard for long. This removes most unsafe Rust and also avoids
leaking Mutex even if the Python world forgets to call `release`.

The only remaining "unsafe" is to bypass a rust-cpython API restriction that is
over strict in this use-case.

Reviewed By: markbt

Differential Revision: D11729852

fbshipit-source-id: 9c914f0769e6903e390982c7aa423bb7e3ae2742