prioritize mononoke tree fetches over local treepacks
Summary:
While this doesn't deliver any gains immediately, it is desirable
because the shortest path to getting trees with pre-populated size and
content sha1 information is via the mononoke api server fetch.

If we use the local trees we will populate the basic metadata slightly
faster but will then pay more to fetch and hash the data.

The calculus is a little more complex really, as having the contents of
a dir locally in a local datapack is probably faster, but we don't have
a good way to know if that is the case right now.

Reviewed By: strager

Differential Revision: D12813490

fbshipit-source-id: 545b90bf20940dab193bef8120712dd9227033e1