restore part of the mercurial.sparse module
Summary:
D12811551 removed the mercurial.sparse module.  However, actual removal of the
module needs to wait until the changes to the eden dirstate extension get
deployed.

This restores the module until all of the Eden changes have been deployed.
This module only contains a dummy `matcher()` function, which is the only part
of the module that Eden referenced.

This was causing exceptions for users with the a mercurial release from today,
but an Eden release from yesterday.  The exception would normally only occur
when using chg, since chg disables lazyimport.  Without chg, lazyimport would
allow the import to succeed even though the module does not exist.

Reviewed By: DurhamG

Differential Revision: D12877889

fbshipit-source-id: 2c6afd9bcb7031727b40fd32fbc1ba1356eb6e9b