mononoke: graceful shutdown
Summary:
This diffs add a signal handler for SIGTERM signal. When it's received then a
terminate process flag is set to true. When this flag is set then no new client
connections will be accepted, and server waits until open connections finish.
The connections can take a long time, so ideally an external process should
sent SIGKILL after a timeout.

Note that this change also makes thrift server thread detached. The reason is
because making it gracefully stop is non-trivial, so for making it detached
should be fine

Reviewed By: farnz

Differential Revision: D12857453

fbshipit-source-id: 6a8f890ff529d74c21fc0c62e16951dd95a3f101