i18n: support ngettext for pluralized strings
Summary:
Support pluralized translation strings using ngettext.

This allows strings that are appropriately pluralized based on the count of the
item.  To use import `_n` from the `i18n` module and provide it with singular
and plural messages, along with the count of the item that should be pluralized:

```
from mercurial.i18n import _n

ui.write(_n("%d item processed", "%d items processed", count) % count)
```

When using `%`-based string formatting in Python, both variants of the format
string must have the same number of subsitutions: it's not possible to leave
out the `%d` in the singular case.

Reviewed By: quark-zju

Differential Revision: D12921684

fbshipit-source-id: 756d1350a827d0451a07279f6884ee57dba6ac9f