mononoke: add a function to extract Phabricator tag
Summary:
In our hooks we may need to get access to test plan, list of reviewers, summary etc.
Lua doesn't have regex, only patterns. So parsing this information in Lua is a bit tricky.
What's even worse is the fact that we can't add any unittests, so debugging and
fixing is harder. Instead let's add a rust function that would do it for us

Reviewed By: farnz

Differential Revision: D10461203

fbshipit-source-id: 328b76171d2c45ceaf5b2a1f5e1d0bd4c326c223