mononoke: add mem write version of bonsai-hg-mapping
Summary:
We'd like to make it so that it's possible to create BlobRepo that can only do
writes in memory. Main motivation for that is to speed up hg changeset
generation from bonsai changeset. We'd like to create all hg changesets
sequentially, but store all the blobs/db entries in memory. Then we can send
blobs to blobstore/rows to database in parallel.

At the moment we only memory-write blobstore, this diff adds the same for
bonsai hg mapping. Note that there is a method that returns all inserts in the
same order as they were made. This is necessary because we need to ensure that
ancestor parents are recorded before descendatns.

Reviewed By: farnz

Differential Revision: D10427432

fbshipit-source-id: a231c5b04734e6efeec207820d2a5b2dbef02d98