remotefilelog: remove writestore concept from union stores
Summary:
Previously, remotefilelog tried to hide which store should receive
writes behind the union store abstraction. This is starting to make things a
little complicated though. For instance, it means that right now both local data
and local history are written via one api on the remotefilelogcontentstore, even
though the remotefilelogcontentstore should only be about data. It also means
that remotefilelogcontentstore had to become aware of pack files, which it can't
read, so it no longer upholds the guarantee that anything written to it can be
immediately read. Overall it's just confusing.

This diff rips out the writestore concept, and instead has the store setup logic
store the writable stores directly on the repo, and changes the write path to
write to those stores directly. Thus removing the notion of pack files from
remotefilelogcontentstore.

A future diff will clean this up even further, and fix the bug where you can't
read data that was just written to a local pack.

Reviewed By: quark-zju

Differential Revision: D12959502

fbshipit-source-id: 85c39c0696febd0972a21f22f3640fd6954901c1