remotefilelog: move file stores onto fileslog
Summary:
Now that all the file store creation logic is inside the fileslog
abstraction, let's also move the storage of all the stores onto this instance,
instead of having them hanging off the repo object.

This matches the manifestlog pattern and gives us better control over the
lifetime of all the filelog stores.

Reviewed By: quark-zju

Differential Revision: D12959500

fbshipit-source-id: f07f8b52bb83a7837e6dc02664bec6111df7a421