Rearrabge buffering in blobimport to speed it up
Summary:
We can't do large blobimports successfully right now, because we
either overload XDB or run out of memory. Move the buffering into a single file
for clarity, and rearrange it so that we have 100 commits worth of data blobs
being uploaded in parallel, and just 10 commits being uploaded in parallel to
keep XDB load down.

Note that the previous buffering was effectively broken - it just buffered the
read from local flash, which is fast, and did not buffer the entry upload

Reviewed By: HarveyHunt

Differential Revision: D12957290

fbshipit-source-id: 1bd9547f0497000c39dd5d505e821ff4de651c6b