hgsubversion: improve performance with treemanifests
Summary:
hgsubversion has some logic that tries to read the list of files in a
given directory. In a tree-only world, it turns the entire tree into a text
file, then splits it by line and bisects over it.

Let's instead use the walk function, which is optimized for this kind of thing
on trees.

Reviewed By: quark-zju

Differential Revision: D12969497

fbshipit-source-id: ef3af9e0022978d6a4922cbb464bfd14248f5501