don't swallow Python errors in tree iteration
Summary:
During the sequence of events that led up to S169085 we tried an hg
push that resulted in an error that looked like this: "RuntimeError:
std::exception". I believe that error was caused by an uncaught
pyexception, replacing Python's current thread exception with a
useless one.

Catch the pyexception and deal with it appropriately.

Reviewed By: quark-zju

Differential Revision: D13112735

fbshipit-source-id: 301d899543ae95084b890c19b00322e69ded07b2