update `eden debug overlay` to handle dir entries being None
Summary:
Update the CLI's overlay parsing code to make sure the `entries` field of
returned `OverlayDir` objects is always set, even if the serialized thrift
data on disk does not have data for this field.

Apparently "default" thrift fields (neither optional nor required) behave
differently in python than in C++.  In C++ these fields are always populated
by deserialization: if they are not present in the serialized data a default
value is used.  In Python it looks like the thrift code simply leaves the
value unmodified (which usually means `None`).  This changes the code to
ensure that this value will default to an empty dictionary if it is not
present.

Reviewed By: chadaustin

Differential Revision: D13135378

fbshipit-source-id: ce5254922e6fc88d6f3d8c58778ad9d45f02741f